--- assets/icons/icons.dart
@@ -1,1 +1,0 @@
-
