+++ lib/screens/home/widgets/contract_section.dart
@@ -1,0 +1,139 @@
+// ignore_for_file: deprecated_member_use
+
+import 'package:defifundr_mobile/core/constants/assets.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+
+class ContractsSection extends StatelessWidget {
+  const ContractsSection({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        const Text(
+          'Contracts',
+          style: TextStyle(
+            fontSize: 14,
+            fontWeight: FontWeight.bold,
+            color: Color(0xFF212121),
+          ),
+        ),
+        const SizedBox(height: 16),
+        Container(
+          padding: const EdgeInsets.all(24),
+          decoration: BoxDecoration(
+            color: Colors.white,
+            borderRadius: BorderRadius.circular(16),
+          ),
+          child: Column(
+            children: [
+              Row(
+                children: [
+                  Container(
+                    height: 40,
+                    width: 40,
+                    decoration: BoxDecoration(
+                        color: Colors.green.withOpacity(0.3), shape: BoxShape.circle),
+                    child: const Icon(
+                      Icons.description_outlined,
+                      color: Color(0xFF2196F3),
+                      size: 22,
+                    ),
+                  ),
+                  const SizedBox(width: 12),
+                  const Text(
+                    'Contracts',
+                    style: TextStyle(
+                      fontSize: 16,
+                      fontWeight: FontWeight.w500,
+                      color: Color(0xFF616161),
+                    ),
+                  ),
+                ],
+              ),
+              const SizedBox(height: 32),
+              Center(child: SvgPicture.asset(AppAssets.noTransaction)),
+              const SizedBox(height: 16),
+              const Text(
+                'Contracts will appear here',
+                style: TextStyle(
+                  fontSize: 16,
+                  color: Color(0xFF9E9E9E),
+                ),
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+class ContractsSectionFilled extends StatelessWidget {
+  const ContractsSectionFilled({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        const Text(
+          'Contracts',
+          style: TextStyle(
+            fontSize: 14,
+            fontWeight: FontWeight.bold,
+            color: Color(0xFF212121),
+          ),
+        ),
+        const SizedBox(height: 16),
+        Container(
+          padding: const EdgeInsets.all(24),
+          decoration: BoxDecoration(
+            color: Colors.white,
+            borderRadius: BorderRadius.circular(16),
+          ),
+          child: Column(
+            children: [
+              Row(
+                children: [
+                  Container(
+                    height: 40,
+                    width: 40,
+                    decoration: BoxDecoration(
+                        color: Colors.green.withOpacity(0.3), shape: BoxShape.circle),
+                    child: const Icon(
+                      Icons.description_outlined,
+                      color: Color(0xFF2196F3),
+                      size: 22,
+                    ),
+                  ),
+                  const SizedBox(width: 12),
+                  const Text(
+                    'Contracts',
+                    style: TextStyle(
+                      fontSize: 16,
+                      fontWeight: FontWeight.w500,
+                      color: Color(0xFF616161),
+                    ),
+                  ),
+                ],
+              ),
+              const SizedBox(height: 32),
+              Center(child: SvgPicture.asset(AppAssets.noTransaction)),
+              const SizedBox(height: 16),
+              const Text(
+                'Contracts will appear here',
+                style: TextStyle(
+                  fontSize: 16,
+                  color: Color(0xFF9E9E9E),
+                ),
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
