+++ lib/screens/home/widgets/home_header.dart
@@ -1,0 +1,57 @@
+import 'package:defifundr_mobile/core/constants/assets.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+
+class Header extends StatelessWidget {
+  final String userName;
+  final String greeting;
+
+  const Header({
+    Key? key,
+    required this.userName,
+    required this.greeting,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        Row(
+          children: [
+            Column(
+              children: [
+                Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      'Good $greeting',
+                      style: const TextStyle(
+                        fontSize: 12,
+                        fontWeight: FontWeight.w500,
+                        color: Color(0xFF757575),
+                      ),
+                    ),
+                    Text(
+                      userName,
+                      style: const TextStyle(
+                        fontSize: 14,
+                        fontWeight: FontWeight.bold,
+                        color: Color(0xFF212121),
+                      ),
+                    ),
+                  ],
+                ),
+              ],
+            ),
+            Image.asset(AppAssets.cloud)
+          ],
+        ),
+        GestureDetector(
+          onTap: () {},
+          child: SvgPicture.asset(AppAssets.notification),
+        ),
+      ],
+    );
+  }
+}
